amdgpu: Split fence detection code to header
