+++ src/video_core/amdgpu/liverpool.cpp
@@ -11,12 +11,11 @@
 #include "core/debug_state.h"
 #include "core/libraries/videoout/driver.h"
 #include "core/memory.h"
+#include "video_core/amdgpu/fence_detector.h"
 #include "video_core/amdgpu/liverpool.h"
 #include "video_core/amdgpu/pm4_cmds.h"
 #include "video_core/renderdoc.h"
 #include "video_core/renderer_vulkan/vk_rasterizer.h"
-
-u64 fence_tick;
 
 namespace AmdGpu {
 
@@ -129,8 +128,6 @@ void Liverpool::Process(std::stop_token stoken) {
                 std::scoped_lock lock{queue.m_access};
                 queue.submits.pop();
 
-                // LOG_WARNING(Render, "End command list");
-
                 --num_submits;
                 std::scoped_lock lock2{submit_mutex};
                 submit_cv.notify_all();
@@ -216,142 +213,6 @@ Liverpool::Task Liverpool::ProcessCeUpdate(std::span<const u32> ccb) {
     FIBER_EXIT;
 }
 
-struct LabelWrite {
-    const PM4Header* packet;
-    VAddr label;
-    u64 value;
-};
-using Fences = std::vector<LabelWrite>;
-
-Fences DetectFences(std::span<const u32> cmd) {
-    Fences fences;
-    while (!cmd.empty()) {
-        const auto* header = reinterpret_cast<const PM4Header*>(cmd.data());
-        const u32 type = header->type;
-
-        switch (type) {
-        default:
-            UNREACHABLE_MSG("Wrong PM4 type {}", type);
-        case 0:
-            UNREACHABLE_MSG("Unimplemented PM4 type 0, base reg: {}, size: {}",
-                            header->type0.base.Value(), header->type0.NumWords());
-        case 2:
-            cmd = NextPacket(cmd, 1);
-            break;
-        case 3:
-            const PM4ItOpcode opcode = header->type3.opcode;
-            switch (opcode) {
-            case PM4ItOpcode::EventWriteEos: {
-                const auto* event_eos = reinterpret_cast<const PM4CmdEventWriteEos*>(header);
-                if (event_eos->command == PM4CmdEventWriteEos::Command::SignalFence) {
-                    fences.emplace_back(header, event_eos->Address<VAddr>(),
-                                        event_eos->DataDWord());
-                    // LOG_WARNING(Render, "EventWriteEos label={:#x}, value={:#x}",
-                    // event_eos->Address<VAddr>(), event_eos->DataDWord());
-                }
-                break;
-            }
-            case PM4ItOpcode::EventWriteEop: {
-                const auto* event_eop = reinterpret_cast<const PM4CmdEventWriteEop*>(header);
-                if (event_eop->int_sel != InterruptSelect::None) {
-                    fences.emplace_back(header);
-                    // LOG_WARNING(Render, "EventWriteEop Irq");
-                }
-                if (event_eop->data_sel == DataSelect::Data32Low) {
-                    fences.emplace_back(header, event_eop->Address<VAddr>(),
-                                        event_eop->DataDWord());
-                    // LOG_WARNING(Render, "EventWriteEop label={:#x}, value32={:#x}",
-                    // event_eop->Address<VAddr>(), event_eop->DataDWord());
-                } else if (event_eop->data_sel == DataSelect::Data64) {
-                    fences.emplace_back(header, event_eop->Address<VAddr>(),
-                                        event_eop->DataQWord());
-                    // LOG_WARNING(Render, "EventWriteEop label={:#x}, value64={:#x}",
-                    // event_eop->Address<VAddr>(), event_eop->DataQWord());
-                }
-                break;
-            }
-            case PM4ItOpcode::ReleaseMem: {
-                const auto* release_mem = reinterpret_cast<const PM4CmdReleaseMem*>(header);
-                if (release_mem->data_sel == DataSelect::Data32Low) {
-                    fences.emplace_back(header, release_mem->Address<VAddr>(),
-                                        release_mem->DataDWord());
-                    // LOG_WARNING(Render, "ReleaseMem label={:#x}, value32={:#x}",
-                    // release_mem->Address<VAddr>(), release_mem->DataDWord());
-                } else if (release_mem->data_sel == DataSelect::Data64) {
-                    fences.emplace_back(header, release_mem->Address<VAddr>(),
-                                        release_mem->DataQWord());
-                    // LOG_WARNING(Render, "ReleaseMem label={:#x}, value64={:#x}",
-                    // release_mem->Address<VAddr>(), release_mem->DataQWord());
-                }
-                break;
-            }
-            case PM4ItOpcode::WriteData: {
-                const auto* write_data = reinterpret_cast<const PM4CmdWriteData*>(header);
-                ASSERT(write_data->dst_sel.Value() == 2 || write_data->dst_sel.Value() == 5);
-                const u32 data_size = (header->type3.count.Value() - 2) * 4;
-                if (data_size <= sizeof(u64) && write_data->wr_confirm) {
-                    u64 value{};
-                    std::memcpy(&value, write_data->data, data_size);
-                    fences.emplace_back(header, write_data->Address<VAddr>(), value);
-                    // LOG_WARNING(Render, "WriteData label={:#x}, value={:#x}",
-                    // write_data->Address<VAddr>(), write_data->data[0]);
-                }
-                break;
-            }
-            case PM4ItOpcode::WaitRegMem: {
-                const auto* wait_reg_mem = reinterpret_cast<const PM4CmdWaitRegMem*>(header);
-                if (wait_reg_mem->mem_space == PM4CmdWaitRegMem::MemSpace::Register) {
-                    break;
-                }
-                const VAddr wait_addr = wait_reg_mem->Address<VAddr>();
-                using Function = PM4CmdWaitRegMem::Function;
-                const u32 mask = wait_reg_mem->mask;
-                const u32 ref = wait_reg_mem->ref;
-                // LOG_WARNING(Render, "WaitRegMem label={:#x}, ref={:#x}, function={}",
-                // wait_reg_mem->Address<VAddr>(), wait_reg_mem->ref,
-                // u32(wait_reg_mem->function.Value()));
-                std::erase_if(fences, [&](const LabelWrite& write) {
-                    if (wait_addr != write.label) {
-                        return false;
-                    }
-                    const u32 value = static_cast<u32>(write.value);
-                    const bool matched = [&] {
-                        switch (wait_reg_mem->function.Value()) {
-                        case Function::LessThan:
-                            return (value & mask) < ref;
-                        case Function::LessThanEqual:
-                            return (value & mask) <= ref;
-                        case Function::Equal:
-                            return (value & mask) == ref;
-                        case Function::NotEqual:
-                            return (value & mask) != ref;
-                        case Function::GreaterThanEqual:
-                            return (value & mask) >= ref;
-                        case Function::GreaterThan:
-                            return (value & mask) > ref;
-                        default:
-                            UNREACHABLE();
-                        }
-                    }();
-                    if (matched) {
-                        // LOG_WARNING(Render, "WaitRegMem Matched with label write");
-                    }
-                    return matched;
-                });
-                break;
-            }
-            default:
-                break;
-            }
-            cmd = NextPacket(cmd, header->type3.NumWords() + 1);
-            break;
-        }
-    }
-
-    // LOG_WARNING(Render, "Final num fences {}", fences.size());
-    return fences;
-}
-
 Liverpool::Task Liverpool::ProcessGraphics(std::span<const u32> dcb, std::span<const u32> ccb) {
     FIBER_ENTER(dcb_task_name);
 
@@ -367,8 +228,7 @@ Liverpool::Task Liverpool::ProcessGraphics(std::span<const u32> dcb, std::span<c
         RESUME_GFX(ce_task);
     }
 
-    // LOG_ERROR(Render, "Graphics");
-    const auto fences = DetectFences(dcb);
+    const auto fence_detector = FenceDetector(dcb);
 
     const auto base_addr = reinterpret_cast<uintptr_t>(dcb.data());
     while (!dcb.empty()) {
@@ -760,11 +620,7 @@ Liverpool::Task Liverpool::ProcessGraphics(std::span<const u32> dcb, std::span<c
                         const u32 value = rasterizer->ReadDataFromGds(event_eos->gds_index);
                         *event_eos->Address() = value;
                     }
-                } else if (std::ranges::contains(fences, header, &LabelWrite::packet) &&
-                           rasterizer) {
-                    ++fence_tick;
-                    // LOG_WARNING(Render, "EventWriteEos label={:#x} fence_tick = {}",
-                    // event_eos->Address<VAddr>(), fence_tick);
+                } else if (fence_detector.IsFence(header) && rasterizer) {
                     rasterizer->CommitPendingGpuRanges();
                 }
                 event_eos->SignalFence([](void* address, u64 data, u32 num_bytes) {
@@ -777,10 +633,7 @@ Liverpool::Task Liverpool::ProcessGraphics(std::span<const u32> dcb, std::span<c
             }
             case PM4ItOpcode::EventWriteEop: {
                 const auto* event_eop = reinterpret_cast<const PM4CmdEventWriteEop*>(header);
-                if (std::ranges::contains(fences, header, &LabelWrite::packet) && rasterizer) {
-                    ++fence_tick;
-                    // LOG_WARNING(Render, "EventWriteEop label={:#x} fence_tick = {}",
-                    // event_eop->Address<VAddr>(), fence_tick);
+                if (fence_detector.IsFence(header) && rasterizer) {
                     rasterizer->CommitPendingGpuRanges();
                 }
                 event_eop->SignalFence([](void* address, u64 data, u32 num_bytes) {
@@ -831,9 +684,7 @@ Liverpool::Task Liverpool::ProcessGraphics(std::span<const u32> dcb, std::span<c
                 const auto* write_data = reinterpret_cast<const PM4CmdWriteData*>(header);
                 ASSERT(write_data->dst_sel.Value() == 2 || write_data->dst_sel.Value() == 5);
                 const u32 data_size = (header->type3.count.Value() - 2) * 4;
-                if (std::ranges::contains(fences, header, &LabelWrite::packet) && rasterizer) {
-                    ++fence_tick;
-                    // LOG_WARNING(Render, "WriteData fence_tick = {}", fence_tick);
+                if (fence_detector.IsFence(header) && rasterizer) {
                     rasterizer->CommitPendingGpuRanges();
                 }
                 if (!write_data->wr_one_addr.Value()) {
@@ -977,9 +828,7 @@ Liverpool::Task Liverpool::ProcessCompute(std::span<const u32> acb, u32 vqid) {
     FIBER_ENTER(acb_task_name[vqid]);
     auto& queue = asc_queues[{vqid}];
 
-    // LOG_ERROR(Render, "Compute");
-    const auto fences = DetectFences(acb);
-
+    const auto fence_detector = FenceDetector(acb);
     boost::container::small_vector<const PM4Header*, 4> indirect_patches;
 
     auto base_addr = reinterpret_cast<VAddr>(acb.data());
@@ -1105,8 +954,6 @@ Liverpool::Task Liverpool::ProcessCompute(std::span<const u32> acb, u32 vqid) {
                         return;
                     }
                     indirect_patches.push_back(header);
-                    // LOG_ERROR(Render, "Rewind GPU range addr={:#x}, size={:#x}", begin, end -
-                    // begin);
                 });
 
             if (must_flush) {
@@ -1193,9 +1040,7 @@ Liverpool::Task Liverpool::ProcessCompute(std::span<const u32> acb, u32 vqid) {
             const auto* write_data = reinterpret_cast<const PM4CmdWriteData*>(header);
             ASSERT(write_data->dst_sel.Value() == 2 || write_data->dst_sel.Value() == 5);
             const u32 data_size = (header->type3.count.Value() - 2) * 4;
-            if (std::ranges::contains(fences, header, &LabelWrite::packet) && rasterizer) {
-                ++fence_tick;
-                // LOG_WARNING(Render, "Compute WriteData fence_tick = {}", fence_tick);
+            if (fence_detector.IsFence(header) && rasterizer) {
                 rasterizer->CommitPendingGpuRanges();
             }
             if (!write_data->wr_one_addr.Value()) {
@@ -1227,9 +1072,7 @@ Liverpool::Task Liverpool::ProcessCompute(std::span<const u32> acb, u32 vqid) {
         }
         case PM4ItOpcode::ReleaseMem: {
             const auto* release_mem = reinterpret_cast<const PM4CmdReleaseMem*>(header);
-            if (std::ranges::contains(fences, header, &LabelWrite::packet) && rasterizer) {
-                ++fence_tick;
-                // LOG_WARNING(Render, "ReleaseMem fence_tick = {}", fence_tick);
+            if (fence_detector.IsFence(header) && rasterizer) {
                 rasterizer->CommitPendingGpuRanges();
             }
             release_mem->SignalFence(static_cast<Platform::InterruptId>(queue.pipe_id));
